requirements.txt
streamlit==1.38.0
pymupdf==1.24.10
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
numpy==1.26.4
pydantic==2.8.2
python-dotenv==1.0.1
sqlalchemy==2.0.34
